NMS-12874: Renamed RemotePoller to PerspectivePoller
