+++ core/upgrade/src/test/resources/NMS-12684/service-configuration.xml-3
@@ -307,7 +307,7 @@ maintained
         <invoke at="stop" pass="0" method="stop"/>
     </service>
     <service>
-        <name>OpenNMS:Name=RemotePollerNG</name>
+        <name>OpenNMS:Name=PerspectivePoller</name>
         <class-name>org.opennms.netmgt.daemon.SimpleSpringContextJmxServiceDaemon</class-name>
         <attribute>
           <name>LoggingPrefix</name>
