+++ features/perspectivepoller/src/main/resources/beanRefContext.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<beans xmlns="http://www.springframework.org/schema/beans" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xsi:schemaLocation="http://www.springframework.org/schema/beans
+        http://www.springframework.org/schema/beans/spring-beans-4.2.xsd" default-lazy-init="true">
+
+    <bean id="perspectivepollerdContext" class="org.springframework.context.support.ClassPathXmlApplicationContext">
+     <constructor-arg>
+       <list>
+         <value>META-INF/opennms/applicationContext-perspectivePollerDaemon.xml</value>
+       </list>
+     </constructor-arg>
+     <constructor-arg ref="daoContext" />
+   </bean>
+
+</beans>
