+++ features/perspectivepoller/src/test/resources/META-INF/opennms/applicationContext-testPerspectivePollerDaemon.xml
