+++ features/perspectivepoller/src/test/resources/poller-configuration-1.xml
