+++ features/perspectivepoller/src/test/resources/poller-configuration-2.xml
