+++ features/perspectivepoller/src/test/resources/threshd-configuration.xml
@@ -1,0 +1,12 @@
+<?xml version="1.0"?>
+<threshd-configuration threads="5">
+
+    <package name="perspectivepoller-thresholding-test">
+        <filter>IPADDR != '0.0.0.0'</filter>
+        <service name="ICMP" interval="30000" user-defined="false" status="on">
+            <parameter key="thresholding-group" value="ICMP" />
+        </service>
+        <outage-calendar>zzz from poll-outages.xml zzz</outage-calendar>
+    </package>
+
+</threshd-configuration>
