+++ features/perspectivepoller/src/test/resources/thresholds.xml
