--- features/remotepollerng/daemon/pom.xml
@@ -1,136 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <parent>
-    <groupId>org.opennms.features</groupId>
-    <artifactId>org.opennms.features.remotepollerng</artifactId>
-    <version>27.0.0-SNAPSHOT</version>
-  </parent>
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>org.opennms.features.remotepollerng</groupId>
-  <artifactId>org.opennms.features.remotepollerng.daemon</artifactId>
-  <packaging>bundle</packaging>
-  <name>OpenNMS :: Features :: Remote Poller NG :: Daemon</name>
-
-  <build>
-    <plugins>
-        <plugin>
-            <groupId>org.apache.felix</groupId>
-            <artifactId>maven-bundle-plugin</artifactId>
-            <extensions>true</extensions>
-            <configuration>
-                <instructions>
-                    <Bundle-RequiredExecutionEnvironment>JavaSE-1.8</Bundle-RequiredExecutionEnvironment>
-                    <Bundle-SymbolicName>${project.artifactId}</Bundle-SymbolicName>
-                    <Bundle-Version>${project.version}</Bundle-Version>
-                </instructions>
-            </configuration>
-        </plugin>
-    </plugins>
-  </build>
-
-  <dependencies>
-      <dependency>
-        <groupId>com.google.guava</groupId>
-        <artifactId>guava</artifactId>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms</groupId>
-          <artifactId>opennms-model</artifactId>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms</groupId>
-          <artifactId>opennms-dao</artifactId>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms</groupId>
-          <artifactId>opennms-config</artifactId>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms</groupId>
-          <artifactId>opennms-dao-api</artifactId>
-      </dependency>
-      <dependency>
-          <groupId>org.quartz-scheduler</groupId>
-          <artifactId>quartz</artifactId>
-          <version>${quartzVersion}</version>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms.core.ipc.rpc</groupId>
-          <artifactId>org.opennms.core.ipc.rpc.api</artifactId>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms.core</groupId>
-          <artifactId>org.opennms.core.daemon</artifactId>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms.features.collection</groupId>
-          <artifactId>org.opennms.features.collection.thresholding.api</artifactId>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms.features.collection</groupId>
-          <artifactId>org.opennms.features.collection.thresholding.impl</artifactId>
-          <version>${project.version}</version>
-      </dependency>
-
-      <dependency>
-          <groupId>junit</groupId>
-          <artifactId>junit</artifactId>
-          <scope>test</scope>
-      </dependency>
-      <dependency>
-          <groupId>org.hamcrest</groupId>
-          <artifactId>hamcrest</artifactId>
-          <version>2.1</version>
-          <scope>test</scope>
-      </dependency>
-      <dependency>
-          <groupId>com.spotify</groupId>
-          <artifactId>hamcrest-pojo</artifactId>
-          <version>1.2.0</version>
-          <scope>test</scope>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms.core.test-api</groupId>
-          <artifactId>org.opennms.core.test-api.services</artifactId>
-          <scope>test</scope>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms.core.test-api</groupId>
-          <artifactId>org.opennms.core.test-api.db</artifactId>
-          <scope>test</scope>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms</groupId>
-          <artifactId>opennms-dao-mock</artifactId>
-          <scope>test</scope>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms</groupId>
-          <artifactId>opennms-services</artifactId>
-          <scope>test</scope>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms.core.ipc.rpc</groupId>
-          <artifactId>org.opennms.core.ipc.rpc.mock-impl</artifactId>
-          <version>${project.version}</version>
-          <scope>test</scope>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms.features.poller</groupId>
-          <artifactId>org.opennms.features.poller.client-rpc</artifactId>
-          <scope>test</scope>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms.features.distributed</groupId>
-          <artifactId>org.opennms.features.distributed.kv-store.blob.postgres</artifactId>
-          <version>${project.version}</version>
-          <scope>test</scope>
-      </dependency>
-      <dependency>
-          <groupId>org.opennms.features.distributed</groupId>
-          <artifactId>org.opennms.features.distributed.kv-store.blob.in-memory</artifactId>
-          <version>${project.version}</version>
-          <scope>test</scope>
-      </dependency>
-  </dependencies>
-</project>
