--- features/remotepollerng/daemon/src/main/resources/META-INF/opennms/applicationContext-remotePollerDaemon.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<beans xmlns="http://www.springframework.org/schema/beans"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xmlns:tx="http://www.springframework.org/schema/tx"
-  xmlns:context="http://www.springframework.org/schema/context"
-  xmlns:util="http://www.springframework.org/schema/util"
-  xmlns:onmsgi="http://xmlns.opennms.org/xsd/spring/onms-osgi"
-  xsi:schemaLocation="
-  http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans-4.2.xsd
-  http://www.springframework.org/schema/context http://www.springframework.org/schema/context/spring-context-4.2.xsd
-  http://www.springframework.org/schema/tx http://www.springframework.org/schema/tx/spring-tx-4.2.xsd
-  http://www.springframework.org/schema/util http://www.springframework.org/schema/util/spring-util-4.2.xsd
-  http://xmlns.opennms.org/xsd/spring/onms-osgi http://xmlns.opennms.org/xsd/spring/onms-osgi.xsd
-">
-
-  <context:annotation-config />
-  <tx:annotation-driven/>
-
-  <onmsgi:reference id="sessionUtils" interface="org.opennms.netmgt.dao.api.SessionUtils"/>
-  <onmsgi:reference id="monitoringLocationDao" interface="org.opennms.netmgt.dao.api.MonitoringLocationDao"/>
-  <onmsgi:reference id="pollerConfig" interface="org.opennms.netmgt.config.PollerConfig"/>
-  <onmsgi:reference id="monitoredServiceDao" interface="org.opennms.netmgt.dao.api.MonitoredServiceDao"/>
-  <onmsgi:reference id="locationAwarePollerClient" interface="org.opennms.netmgt.poller.LocationAwarePollerClient"/>
-  <onmsgi:reference id="applicationDao" interface="org.opennms.netmgt.dao.api.ApplicationDao"/>
-  <onmsgi:reference id="collectionAgentFactory" interface="org.opennms.netmgt.collection.api.CollectionAgentFactory"/>
-  <onmsgi:reference id="persisterFactory" interface="org.opennms.netmgt.collection.api.PersisterFactory"/>
-  <onmsgi:reference id="eventForwarder" interface="org.opennms.netmgt.events.api.EventForwarder"/>
-  <onmsgi:reference id="thresholdingService" interface="org.opennms.netmgt.threshd.api.ThresholdingService"/>
-  <onmsgi:reference id="eventDao" interface="org.opennms.netmgt.dao.api.EventDao"/>
-  <onmsgi:reference id="outageDao" interface="org.opennms.netmgt.dao.api.OutageDao"/>
-
-  <bean id="tracker" class="org.opennms.netmgt.remotepollerng.PerspectiveServiceTracker">
-  </bean>
-
-  <bean id="serviceTrackerListener" class="org.opennms.netmgt.events.api.AnnotationBasedEventListenerAdapter">
-    <property name="annotatedListener" ref="tracker" />
-    <property name="eventSubscriptionService" ref="eventSubscriptionService" />
-  </bean>
-
-  <bean id="daemon" class="org.opennms.netmgt.remotepollerng.RemotePollerd">
-  </bean>
-
-  <bean id="daemonListener" class="org.opennms.netmgt.events.api.AnnotationBasedEventListenerAdapter">
-    <property name="annotatedListener" ref="daemon" />
-    <property name="eventSubscriptionService" ref="eventSubscriptionService" />
-  </bean>
-</beans>
