+++ opennms-base-assembly/src/main/filtered/etc/opennms.properties
@@ -259,7 +259,7 @@ java.awt.headless=true
 # org.opennms.netmgt.collectd.DefaultCollectionAgent.loadSnmpDataOnInit=false
 
 # Rate limit for refreshing perspective poller jobs
-# org.opennms.netmgt.remotepollerng.trackerRefreshRateLimit=30000
+# org.opennms.netmgt.perspectivepoller.trackerRefreshRateLimit=30000
 
 ###### REPORTING ######
 opennms.report.template.dir=${install.dir}/etc
