+++ opennms-base-assembly/src/main/filtered/etc/service-configuration.xml
@@ -307,15 +307,15 @@ maintained
     <invoke at="stop" pass="0" method="stop"/>
   </service>
   <service>
-    <name>OpenNMS:Name=RemotePollerNG</name>
+    <name>OpenNMS:Name=PerspectivePoller</name>
     <class-name>org.opennms.netmgt.daemon.SimpleSpringContextJmxServiceDaemon</class-name>
     <attribute>
       <name>LoggingPrefix</name>
-      <value type="java.lang.String">remotepollerd</value>
+      <value type="java.lang.String">perspectivepollerd</value>
     </attribute>
     <attribute>
       <name>SpringContext</name>
-      <value type="java.lang.String">remotepollerdContext</value>
+      <value type="java.lang.String">perspectivepollerdContext</value>
     </attribute>
     <invoke at="start" pass="0" method="init"/>
     <invoke at="start" pass="1" method="start"/>
