+++ opennms-correlation/drools-correlation-engine/src/test/java/org/opennms/netmgt/correlation/drools/CorrelationRulesTestCase.java
@@ -97,11 +97,11 @@ public abstract class CorrelationRulesTestCase {
     }
 
     protected Event createRemoteNodeLostServiceEvent(int nodeId, String ipAddr, String svcName, int locationMonitor) {
-    	return createEvent(EventConstants.REMOTE_NODE_LOST_SERVICE_UEI, nodeId, ipAddr, svcName, locationMonitor);
+    	return createEvent(EventConstants.PERSPECTIVE_NODE_LOST_SERVICE_UEI, nodeId, ipAddr, svcName, locationMonitor);
     }
 
     protected Event createRemoteNodeRegainedServiceEvent(int nodeId, String ipAddr, String svcName, int locationMonitor) {
-    	return createEvent(EventConstants.REMOTE_NODE_REGAINED_SERVICE_UEI, nodeId, ipAddr, svcName, locationMonitor);
+    	return createEvent(EventConstants.PERSPECTIVE_NODE_REGAINED_SERVICE_UEI, nodeId, ipAddr, svcName, locationMonitor);
     }
 
     protected Event createEvent(String uei, int nodeId, String ipAddr, String svcName, int locationMonitor) {
