+++ opennms-doc/guide-development/src/asciidoc/index.adoc
@@ -71,7 +71,7 @@ include::text/rest/nodes.adoc[]
 include::text/rest/notifications.adoc[]
 include::text/rest/outage_timeline.adoc[]
 include::text/rest/outages.adoc[]
-include::text/rest/remotepoller.adoc[]
+include::text/rest/perspectivepoller.adoc[]
 include::text/rest/requisitions.adoc[]
 include::text/rest/resources.adoc[]
 include::text/rest/rtc.adoc[]
